docs(README): Add GitHub community health badge
Indicate the nature of this repository. Use the GitHub icon from Simple
Icons. Configure https://coolors.co/ with the logo color (#181717) as
the primary color, label background color (#555555) as the secondary
color, and text color (#FFFFFF) as the quaternary color. Use the first
recommended tertiary color (#FF6978) for the message background color.