feat(GitHub-Actions): Grant minimum needed scopes
Grant callable workflows the permissions they require in workflow
templates. The Test workflow requires the contents:write and
pull-requests:read scopes, and the Notify Assignee and Notify Reviewers
workflows require the contents:read scope. In the Test workflow,
pre-commit-action uses the contents:write scope to bump the project
version via Commitizen, and slack-templates uses the contents:read and
pull-requests:read scopes to determine the pull request associated with
a push event to main since the push event itself doesn't contain this
information. In the notify workflows, actions/checkout uses the
contents:read scope to check out the calling repository. Granting these
specific permissions also has the effect of reducing the scope of all
unspecified permissions from read to none. Most data of public
repositories can be read without additional permissions, but the
contents:read and pull-requests:read scopes are needed in the callable
workflows when the calling repository is private. Callable workflows can
only be called when granted the permissions they require, so the same
scopes are needed whether the calling repository is public or private.