fix(Renovate): Stop grouping MegaLinter bumps
Only the pre-commit-hooks repository contains multiple MegaLinter
dependencies, and it now groups MegaLinter bumps in its own Renovate
config. Renovate does not currently offer a flexible mechanism by which
a shared config can use chore for bumps of a group in some cases but fix
in others. Every repository has a development dependency on a MegaLinter
Docker image, so excluding the pre-commit-hooks repository, chore will
be correctly used when it's bumped.