fix(Renovate): Add depName "MegaLinter" to group
matchPackageNames currently matches depNames, so add the depName of the
MegaLinter Docker image, namely "MegaLinter," to a list of package names
to group. Renovate plans to change the behavior of matchPackageNames to
match packageNames in the future, so leave the packageName of the
MegaLinter Python flavor, namely "oxsecurity/megalinter-python," in the
list.