fix(MegaLinter): Run ESLint in list_of_files mode
MegaLinter recently added support for arrays to `*_CLI_EXECUTABLE` in
v7.0.0. This allowed us to run the version of ESLint installed in our
project via Yarn in lieu of the version of ESLint in the MegaLinter
Docker image via npm. This avoided many complications that arose from
ESLint being unable to access Yarn dependencies and allowed us to stop
running ESLint in project mode as a workaround.