fix(Renovate): Correct some dependency types
The npm manager considers Yarn v2+ to be of type `packageManager` rather
than `engines`, so follow that convention for asdf and Yarn v1.

Introduce the new `depType` `config` for the centralized MegaLinter and
Renovate configs as well as Yarn plugins. Use the commit scope `config`
when bumping these dependencies. The centralized MegaLinter and Renovate
configs are both governed by the same dependency, namely this repository
itself, so using different types resulted in a duplicate row in the pull
request descriptions generated by Renovate.