fix(Renovate): Prefer GitHub releases to tags
This repository itself, GitHub-hosted runner images, asdf, and the .NET
SDK all publish GitHub releases in addition to tags. Use the releases as
the datasource so that bumps can't precede the crafting of release notes
and to skip versions that are tagged but never released. Some projects
signal that a tagged version was bad by not creating a release for them.
Continue using GitHub tags to bump the dependencies that don't create
GitHub releases, namely Python and the Yarn licenses plugin.