fix(Renovate): Only prepend "v" when missing
We prepend "v" to the "from" version number,
`currentVersion`, for consistency with the "to"
version number, `prettyNewVersion`, because
Renovate doesn't expose a `prettyCurrentVersion`.
However, an upstream change recently caused
`currentVersion` to start with a "v" when bumping
Node.js. Hence, only prepend a "v" to a semantic
version number that doesn't start with "v."